Add shared login helpers and update tests
