+++ auto_test/test_enrollment.py
@@ -1,12 +1,6 @@
 import config
 from selenium.webdriver.common.by import By
-
-
-def login_student(driver, base_url):
-    driver.get(f"{base_url}/login")
-    driver.find_element(By.ID, "email").send_keys(config.STUDENT_EMAIL)
-    driver.find_element(By.ID, "password").send_keys(config.STUDENT_PASSWORD)
-    driver.find_element(By.CSS_SELECTOR, "button[type='submit']").click()
+from helpers import login_student
 
 
 def test_student_enrollment(driver, base_url):
