+++ auto_test/test_subject_crud.py
@@ -1,12 +1,6 @@
 import config
 from selenium.webdriver.common.by import By
-
-
-def login_admin(driver, base_url):
-    driver.get(f"{base_url}/login")
-    driver.find_element(By.ID, "email").send_keys(config.ADMIN_EMAIL)
-    driver.find_element(By.ID, "password").send_keys(config.ADMIN_PASSWORD)
-    driver.find_element(By.CSS_SELECTOR, "button[type='submit']").click()
+from helpers import login_admin
 
 
 def test_subject_crud(driver, base_url):
